requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
conda==4.3.16
distlib==0.3.4
Django==4.0
django-cors-headers==3.10.1
django-filter==21.1
django-restframework==0.0.1
djangorestframework==3.13.1
filelock==3.4.2
idna==3.3
importlib-metadata==4.9.0
Markdown==3.3.6
node==0.9.28
odict==1.8.0
pipenv==2021.11.23
platformdirs==2.4.1
plumber==1.6
pycosat==0.6.3
pyshorteners==1.0.1
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
urllib3==1.26.8
virtualenv==20.11.0
virtualenv-clone==0.5.7
zipp==3.6.0
zope.component==5.0.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.3
